feat: add delay in openMobileMenu to ensure page load stability
- Added a delay in the `openMobileMenu` method using `waitForSelector` and `waitForTimeout` to improve test reliability.
- Adjusted mobile menu tests to ensure elements are fully loaded before interaction.
- Prepared code for pipeline testing.
